feat: implement ChatArea component with chat session management and input handling
