+++ src/components/chat/MainChat.svelte
@@ -1,24 +1,7 @@
 <script lang="ts">
-	import { settings } from '$lib/stores/settings.svelte';
-	import Speech from '$components/chat/input/Speech.svelte';
-	import TextArea from './input/TextArea.svelte';
-	import { ui } from '$lib/stores/ui';
-	import ChatOptions from './ChatOptions.svelte';
-	import { aiState } from '$lib/stores';
-	import DashBoard from '$components/DashBoard.svelte';
-	import { type ChatParameters, chatParametersState } from '$lib/states/chat.svelte';
-	import { Icon } from '@medyll/idae-slotui-svelte';
-	import MessagesList from './MessagesList.svelte';
-	import { chatMetadata } from '$lib/tools/promptSystem';
-	import { connectionTimer } from '$lib/stores/timer.svelte';
-	import type { DBMessage } from '$types/db';
-	import { WollamaApi } from '$lib/db/wollamaApi';
-	import { get } from 'svelte/store';
-	import { ollamaApiMainOptionsParams } from '$lib/stores/ollamaParams';
-	import type { ChatRequest } from 'ollama/browser';
-	import { ChatSessionManager } from '$lib/tools/chatSessionManager';
-	import { OllamaChatMessageRole } from '$types/ollama';
-	import Images from '$components/chat/input/Images.svelte';
+
+	import DashBoard from '$components/DashBoard.svelte'; 
+	import MessagesList from './MessagesList.svelte'; 
 	import ChatArea from './ChatArea.svelte';
 
 	interface MainChatProps {
@@ -28,125 +11,20 @@
 	let { chatPassKey }: MainChatProps = $props();
 
 	let activeChatId: number | undefined = $state();
-	let chatSessionManager: ChatSessionManager = ChatSessionManager.loadSession();
 
-	chatSessionManager.loadFromPathKey(chatPassKey).then((chat) => (activeChatId = chat?.id));
-	chatSessionManager.setParameters(chatParametersState);
-
-	$effect(() => {
-		console.log(chatSessionManager.session);
-		activeChatId = chatSessionManager.sessionId;
-	});
-
-	let placeholder: string = $derived(chatParametersState.voiceListening ? 'Listening...' : 'Message to ai');
-
-	let disableSubmit: boolean = $derived(
-		chatParametersState.prompt.trim() == '' || chatParametersState.isPrompting || $aiState == 'running'
-	);
-
-	async function sendPrompt(sessionManager: ChatSessionManager, chatParams: ChatParameters) {
-		const config = get(settings);
-		const ollamaOptions = get(ollamaApiMainOptionsParams);
-
-		const { systemMessage, previousMessages, userChatMessage } = await sessionManager.buildMessages(chatParams, config);
-
-		let assistantDbMessage: DBMessage;
-		aiState.set('running');
-		await Promise.all(
-			chatParams.models.map(async (model: string) => {
-				assistantDbMessage =
-					assistantDbMessage ??
-					(await sessionManager.createDbMessage(OllamaChatMessageRole.ASSISTANT, {
-						status:     'idle',
-						role:       OllamaChatMessageRole.ASSISTANT,
-						content:    '',
-						model:      model,
-						tool_calls: []
-					}));
-
-				const request: ChatRequest = {
-					model:    model ?? config?.defaultModel,
-					options:  { ...ollamaOptions, temperature: chatParams.temperature },
-					format:   chatParams.format,
-					stream:   true,
-					messages: [systemMessage, ...previousMessages, userChatMessage]
-				} satisfies ChatRequest;
-
-				const senderGenerate = await WollamaApi.chat(request);
-
-				senderGenerate.onStream = (response) => {
-					const target = assistantDbMessage;
-					sessionManager.onMessageStream(assistantDbMessage, response);
-					ui.setAutoScroll(target.chatId, false);
-				};
-
-				/*{abortController
-					: 	AbortController {signal: AbortSignal},
-					doneCallback : () => {},
-					itr : parseJSON {<closed>}}*/
-				senderGenerate.onEnd = (response) => {
-					sessionManager.onMessageDone(assistantDbMessage, response);
-					aiState.set('done');
-					chatMetadata.checkTitle(sessionManager.sessionId);
-				};
-			})
-		);
-
-		ui.setAutoScroll(chatSessionManager?.sessionId, true);
+	function onChatLoaded(id: number, chatPassKey?: string) {
+		activeChatId = id;
+		window.history.replaceState(history.state, '', `/chat/${chatPassKey}`);
 	}
-
-	async function submitHandler(chatParams: ChatParameters) {
-		if (!chatSessionManager.sessionId) {
-			const session = await chatSessionManager.ChatSessionDB.createSession({
-				models:       [...chatParams?.models],
-				systemPrompt: chatParams.promptSystem
-			});
-			window.history.replaceState(history.state, '', `/chat/${session.dbChat.chatPassKey}`);
-		} else {
-			await chatSessionManager.ChatSessionDB.updateSession({
-				models:       [...chatParams?.models],
-				systemPrompt: chatParams.promptSystem
-				/* ollamaBody: ollamaBodyStore, */
-			});
-		}
-		activeChatId = chatSessionManager.sessionId;
-
-		await sendPrompt(chatSessionManager, chatParams).catch((error) => {
-			console.error('Error sending prompt:', error);
-		});
-	}
-
-	function keyPressHandler(e: KeyboardEvent) {
-		if (e.key === 'Enter' && !e.shiftKey) {
-			e.preventDefault();
-			submitHandler($state.snapshot(chatParametersState));
-			chatParametersState.images = undefined;
-			chatParametersState.prompt = '';
-		}
-	}
-
-	// retrieve model, temperature, format
-	$effect(() => {
-		chatParametersState.mode = $settings.request_mode;
-		chatParametersState.models = [$settings.defaultModel];
-	});
+ 
 </script>
 
 {#snippet input()}
 	<div class="application-chat-main">
-		<ChatArea submitHandler={submitHandler} keyPressHandler={keyPressHandler} placeholder={placeholder} />
-		<!--<div class="text-xs text-center theme-bg p-2">{$t('ui.aiCautionMessage')}</div>-->
+		<ChatArea bind:activeChatId {chatPassKey} {onChatLoaded} />
 	</div>
 {/snippet}
-<!-- <div class="absolute right-4"><AgentPick /></div> -->
-<form
-	hidden
-	id="prompt-form"
-	onsubmit={(event) => {
-		event.preventDefault();
-		submitHandler($state.snapshot(chatParametersState));
-	}}
-></form>
+
 <DashBoard showList={Boolean(activeChatId)}>
 	{#snippet home()}
 		{@render input()}
